Split ContextMenu.js
